Fixes #5369 - BREAKING CHANGE - Move IAcceptTarget to Terminal.Gui.Input namespace (#5395)
